index_prefix (#10)
* initial code commit

* Update README.md

* Added pipelines

* picture output added

* Architecture file updates

* modular Changes

* updated the telegram contoller

* plug and play

common route, controller. Now services are as plug and play. Also added App support service which is a service for REAN_SUPPORT

* plug and play

update for plug and play

* Elastic search integrated, naming convention followed and cleaned all the comments

Elastic search integrated, naming convention followed and cleaned all the comments

* resolve merge conflic

* Update dialogflow.response.service.ts

* Update handle.request.service.ts

* Delete dialogflow.response.service.ts

attempt to resolve merge conflict

* Revert "Delete dialogflow.response.service.ts"

This reverts commit 5770a0db234eb43b1bc73d5f6103232918a2c8df.

* filename changes

Due to some unkown  reason file name is changing

* changed file names

* linting issues

linting issues for files in utility folder and whatsapp.message.service.ts solved

* Solved linting issues

* Updated workflows

* updated dialogflow.response.service

updated dialogflow.response.service in the module folder

* Delete dialogflow.response.service.ts

* removed dublication from covid info service

* removed dublication from /common/response.handler

* reduced dublication in setwebhook in whatsapp.message.service

* send user info in slack

send user info in slack

* userstat updated for creating index prefix

* remove dublication from clentController

* Update client.webhook.controller.ts

* remove dublication in client controller

* Update client.webhook.controller.ts

* Update client.webhook.controller.ts

* Updated the Dev workflow
