+++ .github/workflows/dev-ci-cd.yml
@@ -13,8 +13,6 @@ on:
   push:
     branches: [develop]
 
-  workflow_dispatch:
-
 jobs:
   CodeScan-ESLint:
     runs-on: ubuntu-latest
@@ -37,12 +35,6 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-      # - name: Set the dev environment variables
-      #   run: |
-      #       echo "TASK_DEFINTION_NAME=duploservices-dev-health-guru-bot" >> $GITHUB_ENV
-      #       echo "DUPLO_ID=1b25fb90-59d2-4a35-8d69-98e604e7206d" >> $GITHUB_ENV
-      #       echo "SERVICE_NAME=guru-bot" >> $GITHUB_ENV
-
       - name: Checkout
         uses: actions/checkout@v2
 
@@ -66,7 +58,6 @@ jobs:
           echo "Branch: ${{ steps.vars.outputs.branch }}"
           echo "Sha: ${{ steps.vars.outputs.sha_short }}"
           echo "Repo: ${{ steps.vars.outputs.repo_name }}"
-
 
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v1
