+++ src/api/routes/whatsapp.webhook.routes.ts
@@ -1,21 +1,21 @@
 import express from 'express';
 import { Logger } from '../../common/logger';
 import { injectable } from 'tsyringe';
-import { WhatsappWebhookController } from '../controllers/whatsapp.webhook.controller';
+import { ClientWebhookController } from '../controllers/client.webhook.controller';
 
 @injectable()
 export class WhatsappWebhookRoutes{
 
     constructor(private logger?: Logger,
-        private _whatsappWebhookController?: WhatsappWebhookController){
+        private _clientWebhookController?: ClientWebhookController){
         this.logger.log("Inside whatsapp Routes...");
     }
 
     register (app: express.Application) {
         const router = express.Router();
 
-        router.post(`/:client/${process.env.TELEGRAM_BOT_TOKEN}/send`, this._whatsappWebhookController.sendMessage);
-        router.post(`/:client/${process.env.TELEGRAM_BOT_TOKEN}/receive`, this._whatsappWebhookController.receiveMessage);
+        router.post(`/:client/${process.env.TELEGRAM_BOT_TOKEN}/send`, this._clientWebhookController.sendMessage);
+        router.post(`/:client/${process.env.TELEGRAM_BOT_TOKEN}/receive`, this._clientWebhookController.receiveMessage);
 
         app.use('/v1/', router);
     }
