+++ src/services/slack.service.ts
@@ -6,7 +6,13 @@ const slackClient = new WebClient(process.env.SLACK_TOKEN);
 const slackChannel = process.env.SLACK_CHANNEL_ID;
 
 // Slack Service to Send Message to Slack Channel
-export const send_message = async (message, failureReason, params) => {
+
+export const send_message = async (message, failureReason, params, eventObj) => {
+    var sessionString = eventObj.session.lastIndexOf('/');
+    var client_Id = eventObj.session.substring(sessionString + 1);
+
+    console.log("The evnt ojh issssssssssssssssssss", eventObj);
+
     return new Promise(async (resolve, reject) => {
         try {
             Logger.instance().log(`Sending message to Slack Channel: ${message}`);
@@ -35,6 +41,19 @@ export const send_message = async (message, failureReason, params) => {
                 {
                     type : "section",
                     text : {
+
+                        type : "mrkdwn", text : ":incoming_envelope: *User ID*\n```" + client_Id + "```"
+                    }
+                },
+                {
+                    type : "section",
+                    text : {
+                        type : "mrkdwn", text : ":incoming_envelope: *Message from user*\n```" + eventObj.queryResult.queryText + "```"
+                    }
+                },
+                {
+                    type : "section",
+                    text : {
                         type : "mrkdwn", text : `:vertical_traffic_light: Env: ${process.env.ENVIRONMENT.toLocaleLowerCase()}`
                     }
                 }
@@ -46,11 +65,16 @@ export const send_message = async (message, failureReason, params) => {
                 channel : slackChannel
             });
 
+
+            console.log("asfjgbsk;abfds", result);
+
             Logger.instance().log(`Slack Message Sent: ${result.ok}`);
             resolve(true);
 
         } catch (error) {
             Logger.instance().log_error(error.message, 500, "Slack Service Error!");
+
+            console.log("the Error is", error);
             reject(error.message);
         }
     });
